Code parameter preprocessing set in an independent module
