+++ codeparams.h
@@ -1,0 +1,10 @@
+#ifndef CODEPARAMS_H_
+#define CODEPARAMS_H_
+
+int set_switch_from_codeparams(ItmNs::codeparam_t &codeparams);
+
+std::string stream_xml_string(std::string xml_string, std::string ref);
+
+std::string split_string(std::string s, std::string ref);
+
+#endif /* CODEPARAMS_H_ */
